Making secrets required and renaming dash-server
