+++ client/oauth.html
@@ -1,5 +1,5 @@
 <script>
-  const CLIENT_ID = `23b7d82aec29a3a1a2a8`;
+  const CLIENT_ID = '23b7d82aec29a3a1a2a8';
   const REDIRECT_ORIGIN = window.location.origin;
 
   var queryParams = new URLSearchParams(window.location.search);
